requirements.txt
kanboard==1.1.6
python-dotenv==1.0.1
